+++ qamomile/circuit/frontend/qkernel.py
@@ -37,6 +37,7 @@ from qamomile.circuit.frontend.handle.array import ArrayBase, Vector
 from qamomile.circuit.frontend.handle.containers import Dict
 from qamomile.circuit.frontend.handle.primitives import Bit, Float, Handle, UInt
 from qamomile.circuit.frontend.handle.utils import get_size as _get_size
+from qamomile.circuit.frontend.param_validation import _validate_param_handle
 from qamomile.circuit.frontend.tracer import Tracer, get_current_tracer, trace
 from qamomile.circuit.ir.block import Block, BlockKind
 from qamomile.circuit.ir.operation.call_block_ops import CallBlockOperation
@@ -596,13 +597,7 @@ class QKernel(Generic[P, R]):
         # (``'Qubit' object has no attribute 'shape'``) from ``get_size``
         # during call-time specialization, and a quantum-into-classical
         # mismatch silently miscompiles (a ``Qubit`` bound to a ``Float``
-        # parameter emits ``Rx(0.0)``). ``control.py`` only imports this
-        # module under ``TYPE_CHECKING``, so importing its validator lazily
-        # here keeps the frontend import graph acyclic.
-        from qamomile.circuit.frontend.operation.control import (
-            _validate_param_handle,
-        )
-
+        # parameter emits ``Rx(0.0)``).
         for arg_name, arg_value in bound_args.arguments.items():
             declared = self.input_types.get(arg_name)
             if declared is not None and isinstance(arg_value, Handle):
